backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
alembic